Move VPN service installation over to elevation service
Updates code to use that code too (removing from the service).

This also removes installation that could happen on brave://settings/system
This settings install use-case is obsolete now because the services will be
installed once VPN is actually in use.

This use-case would only happen if the person had VPN but it was set to
IKEv2 and they did not have the services installed. For example, a user
install (not a system install).
