+++ browser/brave_vpn/win/brave_vpn_wireguard_service/service/BUILD.gn
@@ -11,8 +11,6 @@ source_set("service") {
   sources = [
     "brave_wireguard_manager.cc",
     "brave_wireguard_manager.h",
-    "install_utils.cc",
-    "install_utils.h",
     "wireguard_service_runner.cc",
     "wireguard_service_runner.h",
     "wireguard_tunnel_service.cc",
