+++ browser/resources/settings/brave_system_page/brave_vpn_browser_proxy.ts
@@ -6,16 +6,11 @@
 import {sendWithPromise} from 'chrome://resources/js/cr.js';
 
 export interface BraveVPNBrowserProxy {
-  registerWireguardService(): Promise<boolean>;
   isWireguardServiceInstalled(): Promise<boolean>;
   isBraveVpnConnected(): Promise<boolean>;
 }
 
 export class BraveVPNBrowserProxyImpl implements BraveVPNBrowserProxy {
-  registerWireguardService () {
-    return sendWithPromise('registerWireguardService');
-  }
-
   isWireguardServiceInstalled () {
     return sendWithPromise('isWireguardServiceInstalled');
   }
