+++ chromium_src/chrome/elevation_service/DEPS
@@ -1,5 +1,6 @@
 include_rules = [
   "+brave/components/brave_vpn/common/buildflags",
   "+brave/components/brave_vpn/common/wireguard/win",
+  "+brave/elevation_service",
   "+brave/installer",
 ]
