+++ chromium_src/chrome/installer/setup/sources.gni
@@ -17,6 +17,7 @@ if (is_win) {
     brave_chromium_src_chrome_installer_setup_deps += [
       "//brave/components/brave_vpn/browser/connection/ikev2/win:ras_utils",
       "//brave/components/brave_vpn/common/wireguard/win",
+      "//brave/elevation_service",
       "//brave/installer/win/util:vpn_utils",
     ]
   }
