+++ elevation_service/BUILD.gn
@@ -1,0 +1,40 @@
+# Copyright (c) 2023 The Brave Authors. All rights reserved.
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this file,
+# You can obtain one at https://mozilla.org/MPL/2.0/.
+
+import("//brave/build/config.gni")
+
+assert(is_win)
+
+source_set("elevation_service") {
+  sources = [
+    "install_utils.cc",
+    "install_utils.h",
+  ]
+
+  public_deps = [ "//brave/components/brave_vpn/common/wireguard/win:brave_wireguard_manager_idl" ]
+  deps = [
+    "//base",
+    "//brave/browser/brave_vpn/win/brave_vpn_wireguard_service/service",
+    "//brave/browser/brave_vpn/win/brave_vpn_wireguard_service/status_tray",
+    "//brave/components/brave_vpn/common/win",
+    "//brave/components/brave_vpn/common/wireguard/win",
+    "//chrome/installer/util:with_no_strings",
+  ]
+  libs = [
+    "Shell32.lib",
+
+    # TODO(bsclifton): needs to be cleaned up
+    # Please see https://github.com/brave/brave-browser/issues/35288
+    #
+    # This lib is also linked in
+    # brave\browser\brave_vpn\win\brave_vpn_wireguard_service\BUILD.gn
+    # It needs to be included or build has linker error w/ `RoOriginateError`
+    #
+    # This might be expected; since service registration WAS in service before
+    # and now is done here in elevation service. If we're good, we can close
+    # that issue and delete this comment.
+    "runtimeobject.lib",
+  ]
+}
