+++ elevation_service/install_utils.cc
@@ -3,7 +3,9 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-#include "brave/browser/brave_vpn/win/brave_vpn_wireguard_service/service/install_utils.h"
+#include "brave/elevation_service/install_utils.h"
+
+#include <windows.h>  // needed for WM_MENUCOMMAND
 
 #include "base/base_paths.h"
 #include "base/command_line.h"
@@ -14,12 +16,17 @@
 #include "base/win/registry.h"
 #include "base/win/windows_types.h"
 #include "brave/browser/brave_vpn/win/brave_vpn_wireguard_service/service/wireguard_tunnel_service.h"
+#include "brave/browser/brave_vpn/win/brave_vpn_wireguard_service/status_tray/brave_vpn_tray_command_ids.h"
+#include "brave/browser/brave_vpn/win/brave_vpn_wireguard_service/status_tray/status_icon/constants.h"
+#include "brave/browser/brave_vpn/win/brave_vpn_wireguard_service/status_tray/status_icon/icon_utils.h"
 #include "brave/components/brave_vpn/common/win/scoped_sc_handle.h"
 #include "brave/components/brave_vpn/common/win/utils.h"
 #include "brave/components/brave_vpn/common/wireguard/win/service_constants.h"
 #include "brave/components/brave_vpn/common/wireguard/win/service_details.h"
 #include "brave/components/brave_vpn/common/wireguard/win/storage_utils.h"
 #include "brave/components/brave_vpn/common/wireguard/win/wireguard_utils_win.h"
+#include "brave/installer/win/util/brave_vpn_helper_constants.h"
+#include "brave/installer/win/util/brave_vpn_helper_utils.h"
 #include "chrome/installer/util/install_service_work_item.h"
 
 namespace brave_vpn {
@@ -57,6 +64,79 @@ void RemoveFromStartup(const std::wstring& value) {
   }
   key.DeleteValue(value.c_str());
 }
+
+// Microsoft-Windows-NetworkProfile
+// fbcfac3f-8459-419f-8e48-1f0b49cdb85e
+constexpr GUID kNetworkProfileGUID = {
+    0xfbcfac3f,
+    0x8459,
+    0x419f,
+    {0x8e, 0x48, 0x1f, 0x0b, 0x49, 0xcd, 0xb8, 0x5e}};
+
+bool SetServiceTriggerForVPNConnection(SC_HANDLE hService,
+                                       const std::wstring& brave_vpn_entry) {
+  std::wstring brave_vpn_entry_with_null(brave_vpn_entry);
+  brave_vpn_entry_with_null += L'\0';
+  // Allocate and set the SERVICE_TRIGGER_SPECIFIC_DATA_ITEM structure
+  SERVICE_TRIGGER_SPECIFIC_DATA_ITEM deviceData = {0};
+  deviceData.dwDataType = SERVICE_TRIGGER_DATA_TYPE_STRING;
+  // Exclude EOL
+  deviceData.cbData = brave_vpn_entry_with_null.size() *
+                      sizeof(brave_vpn_entry_with_null.front());
+  deviceData.pData = (PBYTE)brave_vpn_entry_with_null.c_str();
+  // Allocate and set the SERVICE_TRIGGER structure
+  SERVICE_TRIGGER serviceTrigger = {0};
+  serviceTrigger.dwTriggerType = SERVICE_TRIGGER_TYPE_CUSTOM;
+  serviceTrigger.dwAction = SERVICE_TRIGGER_ACTION_SERVICE_START;
+  serviceTrigger.pTriggerSubtype = const_cast<GUID*>(&kNetworkProfileGUID);
+  serviceTrigger.cDataItems = 1;
+  serviceTrigger.pDataItems = &deviceData;
+
+  // Allocate and set the SERVICE_TRIGGER_INFO structure
+  SERVICE_TRIGGER_INFO serviceTriggerInfo = {0};
+  serviceTriggerInfo.cTriggers = 1;
+  serviceTriggerInfo.pTriggers = &serviceTrigger;
+
+  // Call ChangeServiceConfig2 with the SERVICE_CONFIG_TRIGGER_INFO level
+  // and pass to it the address of the SERVICE_TRIGGER_INFO structure
+  return ChangeServiceConfig2(hService, SERVICE_CONFIG_TRIGGER_INFO,
+                              &serviceTriggerInfo);
+}
+
+bool ConfigureServiceAutoRestart(const std::wstring& service_name,
+                                 const std::wstring& brave_vpn_entry) {
+  ScopedScHandle scm(::OpenSCManager(nullptr, nullptr, SC_MANAGER_CONNECT));
+  if (!scm.IsValid()) {
+    VLOG(1) << "::OpenSCManager failed. service_name: " << service_name
+            << ", error: " << std::hex << HRESULTFromLastError();
+    return false;
+  }
+  ScopedScHandle service(
+      ::OpenService(scm.Get(), service_name.c_str(), SERVICE_ALL_ACCESS));
+  if (!service.IsValid()) {
+    VLOG(1) << "::OpenService failed. service_name: " << service_name
+            << ", error: " << std::hex << HRESULTFromLastError();
+    return false;
+  }
+
+  if (!brave_vpn::SetServiceFailureActions(service.Get())) {
+    VLOG(1) << "SetServiceFailureActions failed:" << std::hex
+            << HRESULTFromLastError();
+    return false;
+  }
+  if (!SetServiceTriggerForVPNConnection(service.Get(), brave_vpn_entry)) {
+    VLOG(1) << "SetServiceTriggerForVPNConnection failed:" << std::hex
+            << HRESULTFromLastError();
+    return false;
+  }
+  return true;
+}
+
+base::FilePath GetBraveVpnHelperServicePath() {
+  base::FilePath asset_dir = base::PathService::CheckedGet(base::DIR_ASSETS);
+  return asset_dir.Append(brave_vpn::kBraveVPNHelperExecutable);
+}
+
 }  // namespace
 
 bool ConfigureBraveWireguardService(const std::wstring& service_name) {
@@ -138,6 +218,35 @@ bool UninstallBraveWireguardService() {
     return false;
   }
   return true;
+}
+
+bool UninstallStatusTrayIcon() {
+  auto* hWnd = GetBraveVpnStatusTrayIconHWND();
+  if (!hWnd) {
+    return true;
+  }
+
+  return SendMessage(hWnd,
+                     RegisterWindowMessage(kBraveVpnStatusTrayMessageName),
+                     IDC_BRAVE_VPN_TRAY_EXIT, 0) == TRUE;
+}
+
+bool InstallBraveVPNHelperService() {
+  base::CommandLine service_cmd(GetBraveVpnHelperServicePath());
+  installer::InstallServiceWorkItem install_service_work_item(
+      brave_vpn::GetBraveVpnHelperServiceName(),
+      brave_vpn::GetBraveVpnHelperServiceDisplayName(), SERVICE_DEMAND_START,
+      service_cmd, base::CommandLine(base::CommandLine::NO_PROGRAM),
+      brave_vpn::kBraveVpnHelperRegistryStoragePath, {}, {});
+  install_service_work_item.set_best_effort(true);
+  install_service_work_item.set_rollback_enabled(false);
+  if (install_service_work_item.Do()) {
+    auto success =
+        ConfigureServiceAutoRestart(brave_vpn::GetBraveVpnHelperServiceName(),
+                                    brave_vpn::GetBraveVPNConnectionName());
+    return success;
+  }
+  return false;
 }
 
 }  // namespace brave_vpn
