+++ elevation_service/install_utils.h
@@ -3,8 +3,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-#ifndef BRAVE_BROWSER_BRAVE_VPN_WIN_BRAVE_VPN_WIREGUARD_SERVICE_SERVICE_INSTALL_UTILS_H_
-#define BRAVE_BROWSER_BRAVE_VPN_WIN_BRAVE_VPN_WIREGUARD_SERVICE_SERVICE_INSTALL_UTILS_H_
+#ifndef BRAVE_ELEVATION_SERVICE_INSTALL_UTILS_H_
+#define BRAVE_ELEVATION_SERVICE_INSTALL_UTILS_H_
 
 #include <string>
 
@@ -13,7 +13,9 @@ namespace brave_vpn {
 bool ConfigureBraveWireguardService(const std::wstring& service_name);
 bool InstallBraveWireguardService();
 bool UninstallBraveWireguardService();
+bool UninstallStatusTrayIcon();
+bool InstallBraveVPNHelperService();
 
 }  // namespace brave_vpn
 
-#endif  // BRAVE_BROWSER_BRAVE_VPN_WIN_BRAVE_VPN_WIREGUARD_SERVICE_SERVICE_INSTALL_UTILS_H_
+#endif  // BRAVE_ELEVATION_SERVICE_INSTALL_UTILS_H_
