+++ elevation_service/sources.gni
@@ -11,6 +11,7 @@ brave_elevation_service_lib_deps =
 if (is_win && enable_brave_vpn) {
   brave_elevation_service_lib_deps += [
     "//brave/components/brave_vpn/common/wireguard/win",
+    "//brave/elevation_service",
     "//brave/installer/win/util:vpn_utils",
   ]
 }
