+++ installer/win/util/brave_vpn_helper_utils.cc
@@ -24,82 +24,6 @@
 
 namespace brave_vpn {
 
-namespace {
-
-// Microsoft-Windows-NetworkProfile
-// fbcfac3f-8459-419f-8e48-1f0b49cdb85e
-constexpr GUID kNetworkProfileGUID = {
-    0xfbcfac3f,
-    0x8459,
-    0x419f,
-    {0x8e, 0x48, 0x1f, 0x0b, 0x49, 0xcd, 0xb8, 0x5e}};
-
-bool SetServiceTriggerForVPNConnection(SC_HANDLE hService,
-                                       const std::wstring& brave_vpn_entry) {
-  std::wstring brave_vpn_entry_with_null(brave_vpn_entry);
-  brave_vpn_entry_with_null += L'\0';
-  // Allocate and set the SERVICE_TRIGGER_SPECIFIC_DATA_ITEM structure
-  SERVICE_TRIGGER_SPECIFIC_DATA_ITEM deviceData = {0};
-  deviceData.dwDataType = SERVICE_TRIGGER_DATA_TYPE_STRING;
-  // Exclude EOL
-  deviceData.cbData = brave_vpn_entry_with_null.size() *
-                      sizeof(brave_vpn_entry_with_null.front());
-  deviceData.pData = (PBYTE)brave_vpn_entry_with_null.c_str();
-  // Allocate and set the SERVICE_TRIGGER structure
-  SERVICE_TRIGGER serviceTrigger = {0};
-  serviceTrigger.dwTriggerType = SERVICE_TRIGGER_TYPE_CUSTOM;
-  serviceTrigger.dwAction = SERVICE_TRIGGER_ACTION_SERVICE_START;
-  serviceTrigger.pTriggerSubtype = const_cast<GUID*>(&kNetworkProfileGUID);
-  serviceTrigger.cDataItems = 1;
-  serviceTrigger.pDataItems = &deviceData;
-
-  // Allocate and set the SERVICE_TRIGGER_INFO structure
-  SERVICE_TRIGGER_INFO serviceTriggerInfo = {0};
-  serviceTriggerInfo.cTriggers = 1;
-  serviceTriggerInfo.pTriggers = &serviceTrigger;
-
-  // Call ChangeServiceConfig2 with the SERVICE_CONFIG_TRIGGER_INFO level
-  // and pass to it the address of the SERVICE_TRIGGER_INFO structure
-  return ChangeServiceConfig2(hService, SERVICE_CONFIG_TRIGGER_INFO,
-                              &serviceTriggerInfo);
-}
-
-}  // namespace
-
-bool ConfigureServiceAutoRestart(const std::wstring& service_name,
-                                 const std::wstring& brave_vpn_entry) {
-  ScopedScHandle scm(::OpenSCManager(nullptr, nullptr, SC_MANAGER_CONNECT));
-  if (!scm.IsValid()) {
-    VLOG(1) << "::OpenSCManager failed. service_name: " << service_name
-            << ", error: " << std::hex << HRESULTFromLastError();
-    return false;
-  }
-  ScopedScHandle service(
-      ::OpenService(scm.Get(), service_name.c_str(), SERVICE_ALL_ACCESS));
-  if (!service.IsValid()) {
-    VLOG(1) << "::OpenService failed. service_name: " << service_name
-            << ", error: " << std::hex << HRESULTFromLastError();
-    return false;
-  }
-
-  if (!brave_vpn::SetServiceFailureActions(service.Get())) {
-    VLOG(1) << "SetServiceFailureActions failed:" << std::hex
-            << HRESULTFromLastError();
-    return false;
-  }
-  if (!SetServiceTriggerForVPNConnection(service.Get(), brave_vpn_entry)) {
-    VLOG(1) << "SetServiceTriggerForVPNConnection failed:" << std::hex
-            << HRESULTFromLastError();
-    return false;
-  }
-  return true;
-}
-
-base::FilePath GetBraveVpnHelperServicePath() {
-  base::FilePath asset_dir = base::PathService::CheckedGet(base::DIR_ASSETS);
-  return asset_dir.Append(brave_vpn::kBraveVPNHelperExecutable);
-}
-
 // The service starts under system user so we save crashes to
 // %PROGRAMDATA%\BraveSoftware\{service name}\Crashpad
 base::FilePath GetVpnHelperServiceProfileDir() {
@@ -111,24 +35,6 @@ base::FilePath GetVpnHelperServiceProfileDir() {
   return base::FilePath(program_data)
       .Append(install_static::kCompanyPathName)
       .Append(brave_vpn::GetBraveVpnHelperServiceName());
-}
-
-bool InstallBraveVPNHelperService() {
-  base::CommandLine service_cmd(GetBraveVpnHelperServicePath());
-  installer::InstallServiceWorkItem install_service_work_item(
-      brave_vpn::GetBraveVpnHelperServiceName(),
-      brave_vpn::GetBraveVpnHelperServiceDisplayName(), SERVICE_DEMAND_START,
-      service_cmd, base::CommandLine(base::CommandLine::NO_PROGRAM),
-      brave_vpn::kBraveVpnHelperRegistryStoragePath, {}, {});
-  install_service_work_item.set_best_effort(true);
-  install_service_work_item.set_rollback_enabled(false);
-  if (install_service_work_item.Do()) {
-    auto success =
-        ConfigureServiceAutoRestart(brave_vpn::GetBraveVpnHelperServiceName(),
-                                    brave_vpn::GetBraveVPNConnectionName());
-    return success;
-  }
-  return false;
 }
 
 bool IsBraveVPNHelperServiceInstalled() {
@@ -163,53 +69,6 @@ bool IsNetworkFiltersInstalled() {
     return false;
   }
   return current > 0;
-}
-
-HRESULT InstallBraveVPNHelperServiceImpersonated() {
-  HRESULT hr = ::CoImpersonateClient();
-  if (FAILED(hr)) {
-    return hr;
-  }
-
-  absl::Cleanup cleanup([]() { ::CoRevertToSelf(); });
-
-  auto executable_path = brave_vpn::GetBraveVpnHelperServicePath();
-  base::CommandLine command_line(executable_path);
-  command_line.AppendSwitch(brave_vpn::kBraveVpnHelperInstall);
-
-  base::LaunchOptions options = base::LaunchOptions();
-  options.feedback_cursor_off = true;
-  options.wait = true;
-  base::Process proc = base::LaunchProcess(command_line, options);
-  if (!proc.IsValid()) {
-    return brave_vpn::HRESULTFromLastError();
-  }
-
-  return S_OK;
-}
-
-HRESULT InstallBraveWireGuardServiceImpersonated() {
-  HRESULT hr = ::CoImpersonateClient();
-  if (FAILED(hr)) {
-    return hr;
-  }
-
-  absl::Cleanup cleanup([]() { ::CoRevertToSelf(); });
-
-  auto executable_path = brave_vpn::GetBraveVPNWireguardServiceExecutablePath();
-  base::CommandLine command_line(executable_path);
-  command_line.AppendSwitch(
-      brave_vpn::kBraveVpnWireguardServiceInstallSwitchName);
-
-  base::LaunchOptions options = base::LaunchOptions();
-  options.feedback_cursor_off = true;
-  options.wait = true;
-  base::Process proc = base::LaunchProcess(command_line, options);
-  if (!proc.IsValid()) {
-    return brave_vpn::HRESULTFromLastError();
-  }
-
-  return S_OK;
 }
 
 std::wstring GetBraveVPNConnectionName() {
